# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Clippzi API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /gifts:
    get:
      operationId: listGifts
      tags: [gifts]
      summary: Get all available gifts
      parameters:
        - name: category
          in: query
          schema:
            type: ["string", "null"]
        - name: minPrice
          in: query
          schema:
            type: ["number", "null"]
        - name: maxPrice
          in: query
          schema:
            type: ["number", "null"]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Gift"
  /gifts/{id}:
    get:
      operationId: getGift
      tags: [gifts]
      summary: Get a single gift
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Gift"
  /gifts/send:
    post:
      operationId: sendGift
      tags: [gifts]
      summary: Send a gift to a streamer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GiftSendInput"
      responses:
        "201":
          description: Gift sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GiftTransaction"
  /gifts/transactions:
    get:
      operationId: listGiftTransactions
      tags: [gifts]
      summary: List gift transactions (sent or received)
      parameters:
        - name: userId
          in: query
          schema:
            type: ["integer", "null"]
        - name: streamId
          in: query
          schema:
            type: ["integer", "null"]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GiftTransaction"
  /gifts/leaderboard:
    get:
      operationId: getGiftLeaderboard
      tags: [gifts]
      summary: Top gifters leaderboard
      parameters:
        - name: streamId
          in: query
          schema:
            type: ["integer", "null"]
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GifterRank"

  # ── CLIP SHOP ──
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications for a user
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
        - name: unreadOnly
          in: query
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationReadInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"

  # ── MODERATION ──
  /moderation/reports:
    get:
      operationId: listModerationReports
      tags: [moderation]
      summary: List content moderation reports
      parameters:
        - name: status
          in: query
          schema:
            type: ["string", "null"]
            enum: [pending, reviewed, actioned, dismissed, null]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ModerationReport"
    post:
      operationId: createModerationReport
      tags: [moderation]
      summary: Report content for moderation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModerationReportInput"
      responses:
        "201":
          description: Report created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModerationReport"
  /moderation/analyze:
    post:
      operationId: analyzeContent
      tags: [moderation]
      summary: AI-analyze text for bullying or drug content
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContentAnalysisInput"
      responses:
        "200":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentAnalysisResult"
  /moderation/reports/{id}:
    patch:
      operationId: resolveModerationReport
      tags: [moderation]
      summary: Resolve a moderation report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModerationReportUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModerationReport"
components:
  schemas:
...
    User:
      type: object
      required: [id, username, displayName, createdAt]
      properties:
        id: { type: integer }
        username: { type: string }
        displayName: { type: string }
        bio: { type: ["string", "null"] }
        avatarUrl: { type: ["string", "null"] }
        bannerUrl: { type: ["string", "null"] }
        isVerified: { type: boolean }
        isLive: { type: boolean }
        followerCount: { type: integer }
        followingCount: { type: integer }
        postCount: { type: integer }
        totalViews: { type: integer }
        role: { type: string, enum: [user, streamer, admin] }
        createdAt: { type: string }
...
    Gift:
      type: object
      required: [id, name, emoji, price, category, animationUrl]
      properties:
        id: { type: integer }
        name: { type: string }
        emoji: { type: string }
        price: { type: number }
        category: { type: string }
        animationUrl: { type: ["string", "null"] }
        iconUrl: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        rarity: { type: string, enum: [common, rare, epic, legendary] }
        isActive: { type: boolean }
    GiftTransaction:
      type: object
      required: [id, giftId, senderId, receiverId, amount, streamerShare, platformShare, createdAt]
      properties:
        id: { type: integer }
        giftId: { type: integer }
        gift: { $ref: "#/components/schemas/Gift" }
        senderId: { type: integer }
        sender: { $ref: "#/components/schemas/User" }
        receiverId: { type: integer }
        receiver: { $ref: "#/components/schemas/User" }
        streamId: { type: ["integer", "null"] }
        amount: { type: number }
        streamerShare: { type: number }
        platformShare: { type: number }
        quantity: { type: integer }
        message: { type: ["string", "null"] }
        createdAt: { type: string }
    GiftSendInput:
      type: object
      required: [giftId, senderId, receiverId, quantity]
      properties:
        giftId: { type: integer }
        senderId: { type: integer }
        receiverId: { type: integer }
        streamId: { type: integer }
        quantity: { type: integer }
        message: { type: string }
    GifterRank:
      type: object
      required: [rank, user, totalAmount, giftCount]
      properties:
        rank: { type: integer }
        user: { $ref: "#/components/schemas/User" }
        totalAmount: { type: number }
        giftCount: { type: integer }

    # ── CLIP SHOP ──
...
    Notification:
      type: object
      required: [id, userId, type, message, isRead, createdAt]
      properties:
        id: { type: integer }
        userId: { type: integer }
        type: { type: string, enum: [follow, like, comment, gift, live, shop_order, moderation] }
        message: { type: string }
        isRead: { type: boolean }
        relatedUserId: { type: ["integer", "null"] }
        relatedPostId: { type: ["integer", "null"] }
        relatedStreamId: { type: ["integer", "null"] }
        createdAt: { type: string }
    NotificationReadInput:
      type: object
      required: [isRead]
      properties:
        isRead: { type: boolean }

    # ── MODERATION ──
    ModerationReport:
      type: object
      required: [id, reporterId, contentType, contentId, reason, status, createdAt]
      properties:
        id: { type: integer }
        reporterId: { type: integer }
        reporter: { $ref: "#/components/schemas/User" }
        contentType: { type: string, enum: [post, comment, user, stream] }
        contentId: { type: integer }
        reason: { type: string, enum: [bullying, harassment, drugs, spam, nudity, violence, other] }
        description: { type: ["string", "null"] }
        status: { type: string, enum: [pending, reviewed, actioned, dismissed] }
        aiScore: { type: ["number", "null"] }
        aiFlags:
          type: array
          items: { type: string }
        resolvedAt: { type: ["string", "null"] }
        createdAt: { type: string }
    ModerationReportInput:
      type: object
      required: [reporterId, contentType, contentId, reason]
      properties:
        reporterId: { type: integer }
        contentType: { type: string, enum: [post, comment, user, stream] }
        contentId: { type: integer }
        reason: { type: string, enum: [bullying, harassment, drugs, spam, nudity, violence, other] }
        description: { type: string }
    ModerationReportUpdate:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [reviewed, actioned, dismissed] }
    ContentAnalysisInput:
      type: object
      required: [text]
      properties:
        text: { type: string }
        context: { type: ["string", "null"] }
    ContentAnalysisResult:
      type: object
      required: [isSafe, score, flags]
      properties:
        isSafe: { type: boolean }
        score: { type: number }
        flags:
          type: array
          items: { type: string }
        categories:
          type: object
          properties:
            bullying: { type: number }
            drugUse: { type: number }
            harassment: { type: number }
            spam: { type: number }
        recommendation: { type: string, enum: [allow, warn, block, review] }

    # ── PLATFORM STATS ──
...