# lib/api-zod/src/generated/types/giftSendInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Clippzi API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GiftSendInput {
  giftId: number;
  senderId: number;
  receiverId: number;
  streamId?: number;
  quantity: number;
  message?: string;
}