# lib/api-zod/src/generated/types/moderationReportUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Clippzi API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ModerationReportUpdateStatus } from './moderationReportUpdateStatus';

export interface ModerationReportUpdate {
  status: ModerationReportUpdateStatus;
}